Refactor block click and break handling.
Currently, block clicks and breaks are scheduled to run the following tick
by sending a custom packet to server which performs the respective action.
This causes all sorts of issues with mods such as Sponge that expect the
break and click action to occur the same tick as the player action. In
addition, both the vanilla packet and custom packet are sent each time a player
performs a break or click on a storage block.

This commit aims at refactoring this code entirely to follow vanilla standards
and achieve higher compatibility with other mods especially ones that
provide protection.

Here are the bulk of the changes

1. Instead of sending the 'interactShift' change everytime a player clicks
a block on client, this is now handled via the 'BoolConfigUpdateMessage'
packet. When a player joins the server for the first time, the
'interactShift' value will be sent to the server and stored in memory
until the player disconnects. If the player opens up the config gui and
changes it while in game, the same packet will be sent to server to update
value stored or add it as new.

2. Instead of overriding onBlockClicked just to pass hitVec, rayTracing
will also occur on server-side by using Forge's built-in hooks. This
allows the standard onBlockClicked method to be used and avoids headaches
with future updates based on how it is coded currently.

3. Scheduled tasks have been removed entirely as there is no longer a
reason for them. Once a player clicks or breaks a block, the respective
action will occur immediately. Due to this, the BlockClickMessage and
BlockDestroyMessage classes have been removed.