Fix large extractions on small stored item counts returning incorrect stacks
Or, "Change `adjustStoredItemCount` to match documentation"

On a casual read-through I noticed this function can return a negative number even though its docs say otherwise. The callers of this function seem to expect it will only return a positive remainder, notably in `extractItem` where its result is subtracted from the amount requested for extraction. Subtracting the negative result would lead `extractItem` to return a *larger* stack than what is requested, whenever the amount requested is larger than the amount stored in the drawer.